Move value code into separate module
